rename ReferralNav to ReferralPage
